rename project to pyproject-metadata
