+++ README.md
@@ -1,10 +1,10 @@
-# pep621
+# pyproject-metadata
 
-[![pre-commit.ci status](https://results.pre-commit.ci/badge/github/FFY00/python-pep621/main.svg)](https://results.pre-commit.ci/latest/github/FFY00/python-pep621/main)
-[![checks](https://github.com/FFY00/python-pep621/actions/workflows/checks.yml/badge.svg)](https://github.com/FFY00/python-pep621/actions/workflows/checks.yml)
-[![tests](https://github.com/FFY00/python-pep621/actions/workflows/tests.yml/badge.svg)](https://github.com/FFY00/python-pep621/actions/workflows/tests.yml)
-[![codecov](https://codecov.io/gh/FFY00/python-pep621/branch/main/graph/badge.svg?token=9chBjS1lch)](https://codecov.io/gh/FFY00/python-pep621)
-[![Documentation Status](https://readthedocs.org/projects/pep621/badge/?version=latest)](https://pep621.readthedocs.io/en/latest/?badge=latest)
+[![pre-commit.ci status](https://results.pre-commit.ci/badge/github/FFY00/python-pyproject-metadata/main.svg)](https://results.pre-commit.ci/latest/github/FFY00/python-pyproject-metadata/main)
+[![checks](https://github.com/FFY00/python-pyproject-metadata/actions/workflows/checks.yml/badge.svg)](https://github.com/FFY00/python-pyproject-metadata/actions/workflows/checks.yml)
+[![tests](https://github.com/FFY00/python-pyproject-metadata/actions/workflows/tests.yml/badge.svg)](https://github.com/FFY00/python-pyproject-metadata/actions/workflows/tests.yml)
+[![codecov](https://codecov.io/gh/FFY00/python-pyproject-metadata/branch/main/graph/badge.svg?token=9chBjS1lch)](https://codecov.io/gh/FFY00/python-pyproject-metadata)
+[![Documentation Status](https://readthedocs.org/projects/pyproject-metadata/badge/?version=latest)](https://pyproject-metadata.readthedocs.io/en/latest/?badge=latest)
 
 
 > Dataclass for PEP 621 metadata with support for [core metadata] generation
@@ -19,11 +19,11 @@ file (e.g. `PKG-INFO`).
 
 ## Usage
 
-After [installing `pep621`](https://pypi.org/project/pep621/),
+After [installing `pyproject-metadata`](https://pypi.org/project/pyproject-metadata/),
 you can use it as a library in your scripts and programs:
 
 ```python
-from pep621 import StandardMetadata
+from pytproject_metadata import StandardMetadata
 
 parsed_pyproject = { ... }  # you can use parsers like `tomli` to obtain this dict
 metadata = StandardMetadata.from_pyproject(parsed_pyproject)
