+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = pep621
+name = pyproject-metadata
 version = 0.4.0
 description = PEP 621 metadata parsing
 long_description = file: README.md
@@ -16,12 +16,12 @@ classifiers =
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
 project_urls =
-    homepage = https://github.com/FFY00/python-pep621
-    changelog = https://pep621.readthedocs.io/en/stable/changelog.html
+    homepage = https://github.com/FFY00/python-pyproject-metadata
+    changelog = https://pyproject-metadata.readthedocs.io/en/stable/changelog.html
 
 [options]
 py_modules =
-    pep621
+    pyproject_metadata
 install_requires =
     packaging>=19.0
 python_requires = >=3.7
@@ -49,7 +49,7 @@ strict = True
 line_length = 127
 lines_between_types = 1
 lines_after_imports = 2
-known_first_party = pep621
+known_first_party = pyproject-metadata
 
 [coverage:html]
 show_contexts = true
